+++ src/documents/index.js
@@ -1,1 +1,3 @@
 export * from './actor';
+export * from './combat';
+export { createSR4CombatTracker } from './combat-tracker';