+++ src/flows/index.js
@@ -1,2 +1,3 @@
 export * from './spellcasting-flow';
 export * from './defense-flow';
+export * from './defender-picker-flow';