+++ src/models/actor/basecharacter.model.js
@@ -1,62 +1,20 @@
 const fields = foundry.data.fields;
 
 import {
-  realmField,
   conditionMonitorField,
   baseDerivedStatsFields,
+  SR4SheetStatsData,
+  modifiersField,
 } from '@models/shared';
 
 /**
  * @typedef {import('@models/shared').SR4ConditionMonitor} SharedSR4ConditionMonitor
  * @typedef {import('@models/shared').SR4RealmValues} SR4Realms
+ * @typedef {import('@models/shared').SR4SheetStats} SR4SheetStats
+ * @typedef {import('@models/shared').SR4Modifiers} SR4Modifiers
  */
 
-// ---------------------------------------------------------------------------
-// SheetStats
-// ---------------------------------------------------------------------------
-
-/**
- * @typedef {object} SR4SheetStats
- * @property {number} BODY
- * @property {number} CHARISMA
- * @property {number} EDGE
- * @property {number} CURRENTEDGE
- * @property {number} AGILITY
- * @property {number} INTUITION
- * @property {number} ESSENCE
- * @property {number} ASTRALINITIATIVE
- * @property {number} REACTION
- * @property {number} LOGIC
- * @property {number} INITIATIVE
- * @property {number} MATRIXINITIATIVE
- * @property {number} STRENGTH
- * @property {number} WILLPOWER
- * @property {number} MAGIC
- * @property {number} RESONANCE
- */
-
-export class SR4SheetStatsData extends foundry.data.fields.SchemaField {
-  constructor() {
-    super({
-      BODY: new fields.NumberField({ initial: 1, integer: true }),
-      CHARISMA: new fields.NumberField({ initial: 1, integer: true }),
-      EDGE: new fields.NumberField({ initial: 1, integer: true }),
-      CURRENTEDGE: new fields.NumberField({ initial: 1, integer: true }),
-      AGILITY: new fields.NumberField({ initial: 1, integer: true }),
-      INTUITION: new fields.NumberField({ initial: 1, integer: true }),
-      ESSENCE: new fields.NumberField({ initial: 6 }),
-      ASTRALINITIATIVE: new fields.NumberField({ initial: 0, integer: true }),
-      REACTION: new fields.NumberField({ initial: 1, integer: true }),
-      LOGIC: new fields.NumberField({ initial: 1, integer: true }),
-      INITIATIVE: new fields.NumberField({ initial: 2, integer: true }),
-      MATRIXINITIATIVE: new fields.NumberField({ initial: 2, integer: true }),
-      STRENGTH: new fields.NumberField({ initial: 1, integer: true }),
-      WILLPOWER: new fields.NumberField({ initial: 1, integer: true }),
-      MAGIC: new fields.NumberField({ initial: 0, integer: true }),
-      RESONANCE: new fields.NumberField({ initial: 0, integer: true }),
-    });
-  }
-}
+export { SR4SheetStatsData, modifiersField };
 
 // ---------------------------------------------------------------------------
 // Armor
@@ -72,34 +30,6 @@ export const armorField = () =>
   new fields.SchemaField({
     ballistic: new fields.NumberField({ initial: 0, integer: true }),
     impact: new fields.NumberField({ initial: 0, integer: true }),
-  });
-
-// ---------------------------------------------------------------------------
-// Initiative
-// ---------------------------------------------------------------------------
-
-/**
- * @typedef {object} SR4Modifiers
- * @property {{
- *   bonuses: SR4Realms,
- *   passes: SR4Realms
- * }} initiative
- * @property {number} overflowBonus
- * @property {number} woundModBonus
- * @property {number} soakBonus
- * @property {number} generalModifier
- */
-
-export const modifiersField = () =>
-  new foundry.data.fields.SchemaField({
-    initiative: new foundry.data.fields.SchemaField({
-      bonuses: realmField(),
-      passes: realmField(0),
-    }),
-    overflowBonus: new fields.NumberField({ initial: 0, integer: true }),
-    woundModBonus: new fields.NumberField({ initial: 0, integer: true }),
-    generalModifier: new fields.NumberField({ initial: 0, integer: true }),
-    soakBonus: new fields.NumberField({ initial: 0, integer: true }),
   });
 
 // ---------------------------------------------------------------------------