+++ src/models/actor/spirit.model.js
@@ -1,6 +1,10 @@
 const fields = foundry.data.fields;
-import { conditionMonitorField, baseDerivedStatsFields } from '@models/shared';
-import { SR4SheetStatsData } from './basecharacter.model.js';
+import {
+  conditionMonitorField,
+  baseDerivedStatsFields,
+  SR4SheetStatsData,
+  modifiersField,
+} from '@models/shared';
 
 /**
  * @typedef {import('@models/shared').SR4BaseDerivedStats & { stun: number }} SR4SpiritDerivedStats
@@ -12,9 +16,10 @@ import { SR4SheetStatsData } from './basecharacter.model.js';
  * @property {string}  spiritType
  * @property {string}  ownerUuid
  * @property {number}  services
- * @property {import('./basecharacter.model').SR4SheetStats} sheetStats
+ * @property {import('@models/shared').SR4SheetStats} sheetStats
  * @property {SR4SpiritDerivedStats} derivedStats
  * @property {import('@models/shared').SR4ConditionMonitor} conditionMonitor
+ * @property {import('@models/shared').SR4Modifiers} modifiers
  * @property {string}  notes
  */
 
@@ -26,6 +31,7 @@ export class SR4SpiritData extends foundry.abstract.TypeDataModel {
       ownerUuid: new fields.StringField({ initial: '', blank: true }),
       services: new fields.NumberField({ initial: 0, integer: true }),
       sheetStats: new SR4SheetStatsData(),
+      modifiers: modifiersField(),
       derivedStats: new fields.SchemaField({
         ...baseDerivedStatsFields(),
         stun: new fields.NumberField({ initial: 10, integer: true }),