+++ src/utils/dialog/actions/index.js
@@ -1,5 +1,6 @@
 export * from './attack';
 export * from './defense';
+export * from './defender-picker';
 export * from './skills';
 export * from './damage';
 export * from './soak';